This gets tests 4 & 7 working for the most part on all unixes, with some minor difficuties still in the offing (some glitches remain on aix & alpha).
The "SYNC" event handler has been decoupled into one "generator" thread
(which calls waitpid() or poll() in a loop) and a multiplicity of
"handler" threads, which actually do the work of handling various events,
as dispatched to them by the generator.

+++ dyninstAPI/src/BPatch_point.C
@@ -497,7 +497,7 @@ int BPatch_point::getDisplacedInstructionsInt(int maxSize, void* insns)
 
     // So, we return the instruction "overwritten". Wrong, but what the heck...
     const instruction &insn = point->insn();
-    unsigned size = (maxSize < insn.size()) ? maxSize : insn.size();
+    unsigned size = (maxSize < (int)insn.size()) ? maxSize : insn.size();
     memcpy(insns, (const void *)insn.ptr(), size);
     return insn.size();
 }
